fixed comiple errors

+++ acceptance/ArgsparserTestsKeywords.java
@@ -25,9 +25,9 @@ public class ArgsparserTestsKeywords {
 	}
 	
 	public String getProgramOutput(){
-		int l = Integer.valueOf(getLength());
-		int w = Integer.valueOf(getWidth());
-		int h = Integer.valueOf(getHeight());
+		float l = Integer.valueOf(getLength());
+		float w = Integer.valueOf(getWidth());
+		float h = Integer.valueOf(getHeight());
 		
 		int vol = l*w*h;
 		return String.valueOf(vol);
